refactor: move rangepicker logic to decorator
